add NavBarLinks for responsive breakpoints
